+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "files.associations": {
+        "stdlib.h": "c",
+        "get_player.c": "cpp",
+        "personnage.c": "cpp"
+    }
+}
