+++ scenario/personnage.txt
