+++ scenario/story.txt
@@ -1,0 +1,56 @@
+
+    char buf[255];
+    
+    printf("                        🦇\n");
+    printf("                    🦇  pas     🦇\n");
+    printf("                         ^\n");
+    printf("                        || pas la tour de peach\n");
+    printf("                        ||\n");
+    printf("                       /||\\\n");
+    printf("                     __||||__\n");
+    printf("                    |  ||||  |\n");
+    printf("                    |  ||||  |\n");
+    printf("                    |__||||__|\n");
+    printf("                    |  ||||  |\n");
+    printf("                    |  ||||  |\n");
+    printf("                 ___|__||||__|___\n");
+    printf("                |     ||||      |\n");
+    printf("                |     ||||      |\n");
+    printf("                |_____|  |______|\n");
+    printf("                |     ||||      |\n");
+    printf("                |     ||||      |\n");
+    printf("             ___|_____|  |______|___\n");
+    printf("             |        ||||          |\n");
+    printf("             |        ||||          |\n");
+    printf("             |________|  |__________|\n");
+    printf("             |        ||||          |\n");
+    printf("             |        ||||          |\n");
+    printf("         ____|________|  |__________|____\n");
+    printf("          |           ||||              |\n");
+    printf("          |           ||||              |\n");
+    printf("          |___________|  |______________|\n");
+    printf("          |           ||||              |\n");
+    printf("          |           ||||              |\n");
+    printf("      ____|___________|  |______________|____\n");
+    printf("           🦇          🦇       🦇\n");
+    printf("               🦇             🦇\n");
+
+
+
+    printf("\n");
+    printf("**********************************************\n");
+    printf("* Bienvenue dans la Tour des Péchés          *\n");
+    printf("**********************************************\n");
+    printf("\n");
+    printf("Vous vous tenez face à la Tour des Péchés, une imposante structure qui s'élève vers un ciel sombre et orageux.\n");
+    printf("La pierre noire de la tour semble absorber toute lumière, et l'air est lourd, chargé d'un étrange parfum de\n");
+    printf("souffre et de désespoir. Les gargouilles au sommet vous observent silencieusement, et des murmures lointains\n");
+    printf("parviennent à vos oreilles comme des échos de âmes perdues.\n");
+    printf("\n");
+    printf("Il n'y a plus de retour en arrière. En franchissant ces portes, vous vous aventurez dans l'inconnu, prêt à\n");
+    printf("affronter des dangers insoupçonnés et à affronter vos propres démons.\n");
+    printf("\n");
+    printf("Tremblez, car chaque étage vous rapprochera de l'inéluctable, de l'ultime jugement...\n");
+    printf("\n");
+    printf("Êtes-vous prêt à entrer dans la Tour des Péchés ?\n");
+    printf("\n");
