+++ scenario/towerofsins.txt
