+++ src/global.h
@@ -1,0 +1,12 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include <stdbool.h>
+#include <time.h>
+
+#include "./personnages/personnage.c"
+#include "./personnages/get_player.c"
+#include "./personnages/save_player.c"
+#include "./personnages/ennemies/garde_des_limbes.c"
+#include "./combat/combat.c"
