+++ src/main.c
@@ -1,0 +1,15 @@
+#include "global.h"
+
+int main() {
+    personnage player = get_player();
+
+    // tell story
+
+
+    // boucle evenement
+
+
+
+
+    return 0;
+}
